# TinadecCore/tests/__snapshots__/openapi.core.json
{
  "components": {
    "schemas": {
...
      "TinaChatConversationDto": {
        "properties": {
          "accepted_intent_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "allow_cross_workspace": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "last_sequence": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "membership_status": {
            "type": "string"
          },
          "revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "title": {
            "type": "string"
          },
          "workspace_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "title",
          "kind",
          "allow_cross_workspace",
          "revision",
          "last_sequence",
          "accepted_intent_id",
          "membership_status"
        ],
        "type": "object"
      },
      "TinaChatCreateConversationRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "allow_cross_workspace": {
            "default": false,
            "type": "boolean"
          },
          "client_request_id": {
            "type": [
              "null",
              "string"
            ]
          },
          "kind": {
            "default": "group",
            "type": "string"
          },
          "participant_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "title",
          "participant_ids"
        ],
        "type": "object"
      },
      "TinaChatExecuteIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "mode_version_id": {
            "format": "uuid",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "actor_id",
          "mode_version_id"
        ],
        "type": "object"
      },
      "TinaChatExecutionDto": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "intent_id": {
            "format": "uuid",
            "type": "string"
          },
          "participant_id": {
            "format": "uuid",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "session_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "intent_id",
          "participant_id",
          "session_id",
          "run_id",
          "status"
        ],
        "type": "object"
      },
      "TinaChatGenerateIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "audience_participant_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "client_request_id": {
            "type": "string"
          },
          "expected_revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "source_message_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "actor_id",
          "client_request_id",
          "expected_revision",
          "source_message_ids",
          "audience_participant_ids"
        ],
        "type": "object"
      },
...
      "TinaChatIntentDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "expected_revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "actor_id",
          "decision",
          "expected_revision"
        ],
        "type": "object"
      },
      "TinaChatIntentDto": {
        "properties": {
          "author_id": {
            "format": "uuid",
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/TinaChatIntentContent"
          },
          "conversation_id": {
            "format": "uuid",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "decided_by_id": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "source_message_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "author_id",
          "revision",
          "status",
          "content",
          "source_message_ids",
          "decided_by_id",
          "created_at"
        ],
        "type": "object"
      },
      "TinaChatMemberDto": {
        "properties": {
          "joined_after_sequence": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "participant_id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "participant_id",
          "role",
          "status",
          "joined_after_sequence"
        ],
        "type": "object"
      },
      "TinaChatMemberRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string"
          },
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "expected_revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "participant_id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "default": "member",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "participant_id",
          "action",
          "expected_revision"
        ],
        "type": "object"
      },
...
      "TinaChatMessagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TinaChatMessageDto"
            },
            "type": "array"
          },
          "next_cursor": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "type": "object"
      },
...
      "TinaChatProposeIntentRequest": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "format": "uuid",
            "type": "string"
          },
          "audience_participant_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "client_request_id": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/TinaChatIntentContent"
          },
          "expected_revision": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "source_message_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "actor_id",
          "client_request_id",
          "expected_revision",
          "source_message_ids",
          "audience_participant_ids",
          "content"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TinadecCore.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/tina-chat/conversations": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TinaChatConversationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatCreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatConversationDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatConversationDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/intents": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TinaChatIntentDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatProposeIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatIntentDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/intents/generate": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatGenerateIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatIntentDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/intents/{intentId}/decision": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "intentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatIntentDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatIntentDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/intents/{intentId}/execute": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "intentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatExecuteIntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatExecutionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/members": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TinaChatMemberDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatConversationDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/conversations/{id}/messages": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatMessagePage"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatSendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatMessageDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
...
    "/api/v1/tina-chat/participants": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TinaChatParticipantDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatRegisterParticipantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatParticipantDto"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/participants/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatParticipantDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "patch": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatUpdateParticipantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatParticipantDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/participants/{id}/inbox": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_sequence",
            "schema": {
              "format": "int64",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int32",
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatInboxPage"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/participants/{id}/inbox/{messageId}/ack": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
    "/api/v1/tina-chat/workspace-policy": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatWorkspacePolicyDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TinaChatWorkspacePolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TinaChatWorkspacePolicyDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "TinaChat"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}